rework models
