+++ src/aind_slims_api/models/user.py
@@ -1,16 +1,18 @@
-"""Contains a model for a user, and a method for fetching it"""
+"""Contains a model for a user."""
 
 from typing import Optional
 
 from pydantic import Field
 
-from aind_slims_api.core import SlimsBaseModel
+from aind_slims_api.models.base import SlimsBaseModel
 
 
 # TODO: Tighten this up once users are more commonly used
 class SlimsUser(SlimsBaseModel):
     """Model for user information in SLIMS
 
+    Examples
+    --------
     >>> from aind_slims_api.core import SlimsClient
     >>> client = SlimsClient()
     >>> user = client.fetch_model(SlimsUser, username="LKim")
