+++ tests/test_core.py
@@ -12,7 +12,7 @@ from slims.criteria import conjunction, equals
 from slims.internal import Record, _SlimsApiException
 
 from aind_slims_api.core import SlimsAttachment, SlimsClient
-from aind_slims_api.unit import SlimsUnit
+from aind_slims_api.models.unit import SlimsUnit
 
 RESOURCES_DIR = Path(os.path.dirname(os.path.realpath(__file__))) / "resources"
